+++ src/components/spec-document/schema-model/ModelNode.vue
@@ -15,27 +15,15 @@
         :required-fields="resolvedSchemaObject?.required"
       />
     </template>
-
-    <PropertyOneOf
-      v-if="Array.isArray(resolvedSchemaObject?.oneOf) && resolvedSchemaObject?.oneOf?.length"
-      :one-of-list="resolvedSchemaObject?.oneOf"
-    />
-
-    <PropertyAnyOf
-      v-if="Array.isArray(resolvedSchemaObject?.anyOf) && resolvedSchemaObject?.anyOf?.length"
-      :any-of-list="resolvedSchemaObject?.anyOf"
-    />
   </div>
 </template>
 
 <script setup lang="ts">
 import { computed } from 'vue'
 import type { PropType } from 'vue'
+import ModelProperty from './ModelProperty.vue'
 import type { SchemaObject } from '@/types'
 import { isValidSchemaObject, resolveSchemaObjectFields } from '@/utils'
-import ModelProperty from './ModelProperty.vue'
-import PropertyAnyOf from './property-fields/PropertyAnyOf.vue'
-import PropertyOneOf from './property-fields/PropertyOneOf.vue'
 
 const props = defineProps({
   schema: {