+++ CHANGELOG.md
@@ -1,3 +1,53 @@
+## [1.42.2](https://github.com/Kong/spec-renderer/compare/v1.42.1...v1.42.2) (2024-07-09)
+
+
+### Bug Fixes
+
+* **body-content-list:** resolve schema object before trying to remove readonly fields from it ([#194](https://github.com/Kong/spec-renderer/issues/194)) ([971dbb9](https://github.com/Kong/spec-renderer/commit/971dbb972786958dacb021d737071d28dbb2fab9))
+
+## [1.42.1](https://github.com/Kong/spec-renderer/compare/v1.42.0...v1.42.1) (2024-07-09)
+
+
+### Bug Fixes
+
+* use selectdropdown component [KHCP-12513] ([#187](https://github.com/Kong/spec-renderer/issues/187)) ([369da1d](https://github.com/Kong/spec-renderer/commit/369da1d68f78ec00169101d36881b8aed1b9b75f))
+
+# [1.42.0](https://github.com/Kong/spec-renderer/compare/v1.41.0...v1.42.0) (2024-07-09)
+
+
+### Features
+
+* oneof/anyof select-input [KHCP-12371] ([#183](https://github.com/Kong/spec-renderer/issues/183)) ([ef8eff7](https://github.com/Kong/spec-renderer/commit/ef8eff7fd4ee30f86cdd0e510f1cfdeca5cd2d70))
+
+# [1.41.0](https://github.com/Kong/spec-renderer/compare/v1.40.4...v1.41.0) (2024-07-09)
+
+
+### Features
+
+* add support for `allOf` [KHCP-11831] ([#188](https://github.com/Kong/spec-renderer/issues/188)) ([4ec424a](https://github.com/Kong/spec-renderer/commit/4ec424a73b554b32716a75ae8cbc6ead7dc2935d))
+* requiredOnly toggle for TryIt and request sample [KHCP-12452] ([#180](https://github.com/Kong/spec-renderer/issues/180)) ([18dd284](https://github.com/Kong/spec-renderer/commit/18dd28473d9db872716e5c9c290a03c6923d46db))
+
+## [1.40.4](https://github.com/Kong/spec-renderer/compare/v1.40.3...v1.40.4) (2024-07-09)
+
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies with stable version ([#191](https://github.com/Kong/spec-renderer/issues/191)) ([e59b738](https://github.com/Kong/spec-renderer/commit/e59b738273c1b923e93223af8d69e639850d04de))
+
+## [1.40.3](https://github.com/Kong/spec-renderer/compare/v1.40.2...v1.40.3) (2024-07-08)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @kong/icons to ^1.14.2 ([#190](https://github.com/Kong/spec-renderer/issues/190)) ([a89d596](https://github.com/Kong/spec-renderer/commit/a89d5969bc6dbb8875fc71d0437206ec8bed8691))
+
+## [1.40.2](https://github.com/Kong/spec-renderer/compare/v1.40.1...v1.40.2) (2024-07-08)
+
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies with stable version ([#179](https://github.com/Kong/spec-renderer/issues/179)) ([44bb35f](https://github.com/Kong/spec-renderer/commit/44bb35f32b086ef79026f25c479b87d28b7aa82e))
+
 ## [1.40.1](https://github.com/Kong/spec-renderer/compare/v1.40.0...v1.40.1) (2024-07-05)